add merkle contract
