+++ hyperlane/merkle/src/lib.rs
@@ -1,0 +1,5 @@
+mod execute;
+mod query;
+mod state;
+
+pub use {execute::*, query::*, state::*};
