+++ hyperlane/types/src/lib.rs
@@ -3,6 +3,7 @@ pub mod hook;
 pub mod ism;
 pub mod mailbox;
 pub mod merkle;
+pub mod merkle_tree;
 pub mod recipient;
 
 pub use address::*;
